Move Callback type out of Publisher SDK
I want to use it in the Subscriber SDK too.
